spectrum: Use uint32_t instead of uint8_t for m_rbNum and m_timeGranularity in TraceFadingLossModel
With NR we can have a larger number of RBs than we could have
in past with LTE. In LTE the maximum was 100RBs for 20MHz BW,
while in NR it could go e.g. up to 275.
Hence, uint8_t is not sufficient any more.

Similarly, m_timeGranularity needs to support higher values.

+++ src/spectrum/model/trace-fading-loss-model.cc
@@ -84,7 +84,7 @@ TraceFadingLossModel::GetTypeId()
                           "The number of RB the trace is made of (default 100)",
                           UintegerValue(100),
                           MakeUintegerAccessor(&TraceFadingLossModel::m_rbNum),
-                          MakeUintegerChecker<uint8_t>())
+                          MakeUintegerChecker<uint32_t>())
             .AddAttribute(
                 "RngStreamSetSize",
                 "The number of RNG streams reserved for the fading model. The maximum number of "
