+++ src/spectrum/model/trace-fading-loss-model.h
@@ -122,9 +122,9 @@ class TraceFadingLossModel : public SpectrumPropagationLossModel
     Time m_traceLength;               ///< the trace time
     uint32_t m_samplesNum;            ///< number of samples
     Time m_windowSize;                ///< window size
-    uint8_t m_rbNum;                  ///< RB number
+    uint32_t m_rbNum;                 ///< RB number
     mutable Time m_lastWindowUpdate;  ///< time of last window update
-    uint8_t m_timeGranularity;        ///< time granularity
+    uint32_t m_timeGranularity;       ///< time granularity
     mutable uint64_t m_currentStream; ///< the current stream
     mutable uint64_t m_lastStream;    ///< the last stream
     uint64_t m_streamSetSize;         ///< stream set size
